requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.7
Django==4.2.7
django-cors-headers==4.3.0
django-environ==0.10.0
djangorestframework==3.14.0
filelock==3.12.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
platformdirs==3.10.0
psycopg2==2.9.7
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.24.3